Add POM Command.

+++ xn.h
@@ -42,7 +42,7 @@ public:
 	void getLIVersion();
 	void getLIAddress();
 	void setLIAddress(uint8_t addr, CPXnCb ok = nullptr, CPXnCb err = nullptr);
-	void PomWriteCv(LocoAddr, uint8_t cv, uint8_t value, CPXnCb ok = nullptr,
+	void PomWriteCv(LocoAddr, uint16_t cv, uint8_t value, CPXnCb ok = nullptr,
 	                CPXnCb err = nullptr);
 	void setSpeed(LocoAddr, uint8_t speed, bool direction, CPXnCb ok = nullptr,
 	              CPXnCb err = nullptr);
